+++ piskel-clone-app/src/app/app.module.ts
@@ -2,11 +2,11 @@ import {HttpClientModule} from '@angular/common/http';
 import {NgModule} from '@angular/core';
 import {BrowserModule} from '@angular/platform-browser';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
-import {AnimationContainerModule} from '@shared/animation-container/animation-container.module';
-import {CanvasContainerModule} from '@shared/canvas-container/canvas-container.module';
-import {FrameContainerModule} from '@shared/frame-container/frame-container.module';
-import {HeaderContainerModule} from '@shared/header-container/header-container.module';
-import {ToolsModule} from '@shared/tools-container/tools.module';
+import {AnimationContainerModule} from '@components/animation-container/animation-container.module';
+import {CanvasContainerModule} from '@components/canvas-container/canvas-container.module';
+import {FrameContainerModule} from '@components/frame-container/frame-container.module';
+import {HeaderContainerModule} from '@components/header-container/header-container.module';
+import {ToolsModule} from '@components/tools-container/tools.module';
 import {AppRoutingModule} from './app-routing.module';
 import {AppComponent} from './app.component';
