+++ piskel-clone-app/src/app/components/animation-container/animation-container.component.scss
@@ -1,0 +1,39 @@
+@import 'vars';
+@import 'mixins';
+
+.animation-preview {
+  @include block-size(99%, 200px);
+  
+  border: 1px solid red;
+  image-rendering: pixelated;
+  background-color: $white;
+}
+  
+.animation__container {
+  position: relative;
+  width: 100%;
+}
+
+.animation__fullscreen {
+  @include border(transparent);
+  
+  position: absolute;
+  right: 0;
+  cursor: pointer;
+}
+
+.animation__fps {
+  @include flex;
+  background: $white;
+  box-shadow: 0px 4px 4px $shadow-color;
+}
+
+.animation__fps-number {
+  @include flex;
+  @include flex-center;
+
+  width: 55px;
+  margin-right: 5px;
+
+  text-transform: uppercase;
+}
