+++ piskel-clone-app/src/app/components/canvas-container/canvas-container.component.scss
