+++ piskel-clone-app/src/app/components/frame-container/frame-container.component.scss
