+++ piskel-clone-app/src/app/components/frame-container/frames/methods/addFrame.js
