+++ piskel-clone-app/src/app/components/header-container/size-methods/currentSizeCanvas.js
