+++ piskel-clone-app/src/app/components/tools-container/entities/constants.ts
