+++ piskel-clone-app/src/app/components/tools-container/entities/interfaces.ts
