+++ piskel-clone-app/src/app/components/tools-container/tools.component.scss
