--- piskel-clone-app/src/app/shared/animation-container/animation-container.component.html
@@ -1,3 +1,0 @@
-<p>
-  animation-container works!
-</p>
